+++ win32select.c
@@ -126,7 +126,7 @@ do_fd_set(struct win32op *op, struct idx_info *ent, evutil_socket_t s, int read)
 		if (ent->write_pos_plus1 > 0)
 			return (0);
 	}
-	if (set->fd_count == op->fd_setsz) {
+	if ((int)set->fd_count == op->fd_setsz) {
 		if (realloc_fd_sets(op, op->fd_setsz*2))
 			return (-1);
 		/* set pointer will have changed and needs reiniting! */
@@ -155,7 +155,7 @@ do_fd_clear(struct event_base *base,
 	}
 	if (i < 0)
 		return (0);
-	if (--set->fd_count != i) {
+	if (--set->fd_count != (unsigned)i) {
 		struct idx_info *ent2;
 		SOCKET s2;
 		s2 = set->fd_array[i] = set->fd_array[set->fd_count];
